separate district code into 5g and 4g files
